WIP: undo me
This is in the middle of trying to put a stream manager or something
into the process so that when we have a JP2 to look up, we can have an
abstract asset of some kind which can generate any number of streamers.
For filesystem stuff, this would just spit out the results of os.Open
each time a Streamer is requested.  For in-memory streams, it would use
some of the other stuff developed in this branch.  For S3, it would use
calls to the S3 SDK to pull byte ranges, and have something that tracks
where its "file pointer is" - e.g., where it should next read.